requirements.txt
aiohttp==3.5.1
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
dlib==19.16.0
idna==2.8
jikanpy==2.2.0
kiwisolver==1.0.1
matplotlib==3.0.2
multidict==4.5.2
networkx==2.2
numpy==1.15.4
Pillow==5.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyWavelets==1.0.1
requests==2.21.0
scikit-image==0.14.1
scipy==1.2.0
six==1.12.0
toolz==0.9.0
urllib3==1.24.1
yarl==1.3.0
